Refactor Cell class for customizable representations
Adjustments have been made to the Cell class, primarily replacing the return of a static string method in getRepresentation() with a flexible string value. An additional method was created to set the owner's representation as per the TicTacToe game move. This allows for each cell's owner representation to change dynamically as the game progresses.