The provided commit message does not correspond to the code differences shown. Here is an appropriate commit message:
Commit Message:
Simplify BoardManager class and adjust to a fixed board size

Eliminated the flexibility of choosing the board-size by setting a fixed size of 3 in the BoardManager class. Also, removed the 'displayBoard' method as well as some unnecessary comments. This refactoring is in line with changes in the game specifications.